Extract `expect(sut:FeedLoader,toCompleteWith:Result)` test helper method into an extension of `XCTestCase` to reuse across multiple test files
