+++ EssentialApp/EssentialAppTests/Helpers/FeedLoader+XCTestCaseHelpers.swift
@@ -1,0 +1,30 @@
+//
+//  FeedLoader+XCTestCaseHelpers.swift
+//  EssentialAppTests
+//
+//  Created by TCode on 28/5/22.
+//
+
+import EssentialFeed
+import XCTest
+
+extension XCTestCase {
+    func expect(_ sut: FeedLoader, toCompleteWith expectedResult: FeedLoader.Result, file: StaticString = #file, line: UInt = #line) {
+        let exp = expectation(description: "Wait for load completion")
+
+        sut.load { receivedResult in
+            switch (receivedResult, expectedResult) {
+            case let (.success(receivedFeed), .success(expectedFeed)):
+                XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
+            case (.failure, .failure):
+                break
+            default:
+                XCTFail("Expected \(expectedResult), got \(receivedResult) instead")
+            }
+
+            exp.fulfill()
+        }
+
+        wait(for: [exp], timeout: 1.0)
+    }
+}
